created eventTabs its own file and cleaned unnecessary imports and code
