+++ components/event/EventTabs.js
@@ -1,0 +1,174 @@
+import { BiSortAlt2 } from "react-icons/bi";
+import { FaListUl, FaMicrophoneAlt } from "react-icons/fa";
+import { MdOutlineOnlinePrediction, MdOutlinePeople } from "react-icons/md";
+
+export function EventTabs({ tabs, setTabs, sortEvents, setSortEvents }) {
+  const classNames = (...classes) => classes.filter(Boolean).join(" ");
+  const changeTab = (e, value) => {
+    e.preventDefault();
+    console.log(e.target.value);
+    setTabs(
+      tabs.map((tab) =>
+        tab.title === e.target?.value || tab.title === value
+          ? { ...tab, current: true }
+          : { ...tab, current: false }
+      )
+    );
+  };
+
+  const getDataKeyAndSortKey = (tabName) => {
+    let dataKeyObj = {};
+    switch (tabName) {
+      case "cfpOpen":
+        dataKeyObj.dataKey = "cfpOpen";
+        dataKeyObj.sortKey = "date.start";
+        break;
+      case "inPerson":
+        dataKeyObj.dataKey = "inPerson";
+        dataKeyObj.sortKey = "date.start";
+        break;
+      case "virtual":
+        dataKeyObj.dataKey = "virtual";
+        dataKeyObj.sortKey = "date.start";
+        break;
+      default:
+        dataKeyObj.dataKey = "all";
+        dataKeyObj.sortKey = "date.start";
+    }
+    return dataKeyObj;
+  };
+
+  const sortUserTabItems = (tabName, order) => {
+    const { dataKey, sortKey } = getDataKeyAndSortKey(tabName);
+    sortEvents[dataKey].sort(function (a, b) {
+      console.log(sortEvents[dataKey]);
+      const aVal = sortKey.includes(".")
+        ? getNested(a, sortKey.split("."))
+        : a[sortKey];
+      const bVal = sortKey.includes(".")
+        ? getNested(b, sortKey.split("."))
+        : b[sortKey];
+      if (tabName === "My Links") {
+        if (order === "ASC") {
+          return aVal.toLowerCase() > bVal.toLowerCase()
+            ? 1
+            : aVal.toLowerCase() < bVal.toLowerCase()
+            ? -1
+            : 0;
+        } else {
+          return aVal.toLowerCase() < bVal.toLowerCase()
+            ? 1
+            : aVal.toLowerCase() > bVal.toLowerCase()
+            ? -1
+            : 0;
+        }
+      } else {
+        if (order === "ASC") {
+          return new Date(aVal) > new Date(bVal)
+            ? 1
+            : new Date(aVal) < new Date(bVal)
+            ? -1
+            : 0;
+        } else {
+          return new Date(aVal) < new Date(bVal)
+            ? 1
+            : new Date(aVal) > new Date(bVal)
+            ? -1
+            : 0;
+        }
+      }
+    });
+    setSortEvents({ ...sortEvents });
+  };
+
+  const getNested = (obj, args) => {
+    return args.reduce((obj, level) => obj && obj[level], obj);
+  };
+
+  return (
+    <div>
+      <div className="sm:hidden">
+        <label htmlFor="tabs" className="sr-only">
+          Select a tab
+        </label>
+        <select
+          id="tabs"
+          name="tabs"
+          onChange={changeTab}
+          className="block w-full rounded-md border-gray-300 focus:border-indigo-500 focus:ring-indigo-500"
+          defaultValue={tabs.find((tab) => tab.current)?.title}
+        >
+          {tabs.map((tab) => (
+            <option key={tab.key}>{tab.title}</option>
+          ))}
+        </select>
+      </div>
+      <div className="hidden sm:block">
+        <div className="border-b border-gray-200">
+          <nav className="-mb-px flex" aria-label="Tabs">
+            {tabs.map((tab) => (
+              <a
+                key={tab.key}
+                href={tab.href}
+                onClick={(e) => changeTab(e, tab.title)}
+                className={classNames(
+                  tab.current
+                    ? "border-indigo-500 text-indigo-600"
+                    : "border-transparent text-gray-500 hover:text-gray-700 hover:border-gray-300",
+                  "w-1/4 py-4 px-1 text-center border-b-2 font-medium text-sm flex justify-center items-center gap-2"
+                )}
+                aria-current={tab.current ? "page" : undefined}
+              >
+                {tab.key === "all" && (
+                  <FaListUl title="Show All" className="text-base" />
+                )}
+                {tab.key === "virtual" && (
+                  <MdOutlineOnlinePrediction
+                    title="Virtual event"
+                    className="text-xl"
+                  />
+                )}
+                {tab.key === "inPerson" && (
+                  <MdOutlinePeople
+                    title="In person event"
+                    className="text-xl"
+                  />
+                )}
+                {tab.key === "cfpOpen" && (
+                  <FaMicrophoneAlt title="CFP is open" className="text-base" />
+                )}
+                <span>
+                  {tab.title} ({tab.total})
+                </span>
+                {tab.current && (
+                  <BiSortAlt2
+                    size="20"
+                    className="hover:text-gray-400"
+                    onClick={(e) => {
+                      e.preventDefault();
+                      e.stopPropagation();
+                      setTabs(
+                        tabs.map((tab) =>
+                          tab.current
+                            ? {
+                                ...tab,
+                                order: tab.order === "ASC" ? "DESC" : "ASC",
+                              }
+                            : { ...tab }
+                        )
+                      );
+                      sortUserTabItems(
+                        tab.title,
+                        tab.order === "ASC" ? "DESC" : "ASC"
+                      );
+                    }}
+                  />
+                )}
+              </a>
+            ))}
+          </nav>
+        </div>
+      </div>
+    </div>
+  );
+}
