+++ pages/events.js
@@ -1,19 +1,10 @@
 import Head from "next/head";
 import { useState } from "react";
-import { IconContext } from "react-icons";
-import { FaListUl, FaMicrophoneAlt } from "react-icons/fa";
-import {
-  MdOutlineOnlinePrediction,
-  MdOutlinePeople,
-  MdSettingsRemote,
-} from "react-icons/md";
-
 import EventCard from "../components/event/EventCard";
-import Alert from "../components/Alert";
 import Page from "../components/Page";
-import EventKey from "../components/event/EventKey";
-import UserTabs from "../components/user/UserTabs";
-import { BiSortAlt2 } from "react-icons/bi";
+import { EventTabs } from "../components/event/EventTabs";
+import { FaListUl, FaMicrophoneAlt } from "react-icons/fa";
+import { MdOutlineOnlinePrediction, MdOutlinePeople } from "react-icons/md";
 
 export async function getServerSideProps(context) {
   let events = [];
@@ -30,8 +21,6 @@ export async function getServerSideProps(context) {
 }
 
 export default function Events({ events }) {
-  const [eventType, seteventType] = useState("all");
-
   let categorisedEvents = {
     all: events,
     virtual: events.filter((event) => event.isVirtual === true),
@@ -88,8 +77,6 @@ export default function Events({ events }) {
 
   const [tabs, setTabs] = useState(displayTabs);
   const [sortEvents, setSortEvents] = useState(categorisedEvents);
-
-  // console.log(tabs, "bing bing boo");
   return (
     <>
       <Head>
@@ -99,21 +86,7 @@ export default function Events({ events }) {
       </Head>
       <Page>
         <h1 className="text-4xl mb-4 font-bold">Community events</h1>
-        {/* previous code */}
-        {/* <EventKey
-          categorisedEvents={categorisedEvents}
-          onToggleEventType={(newValue) => seteventType(newValue)}
-        /> */}
-
-        {/* {!events.length && <Alert type="info" message="No events found" />}
-        <ul role="list" className="divide-y divide-gray-200">
-          {categorisedEvents[eventType].map((event, index) => (
-            <EventCard event={event} username={event.username} key={index} />
-          ))}
-        </ul> */}
-
-        {/* new code */}
-        <UserTabs2
+        <EventTabs
           tabs={tabs}
           setTabs={setTabs}
           setSortEvents={setSortEvents}
@@ -143,205 +116,7 @@ export default function Events({ events }) {
               </ul>
             )
         )}
-
-        {/* {tabs.find((tab) => tab.key === "all") &&
-          tabs.find((tab) => tab.key === "all").current && (
-            <div className="bg-red-500 p-8 mt-12">
-              <h1>all</h1>
-            </div>
-          )}
-
-        {tabs.find((tab) => tab.key === "cfpOpen") &&
-          tabs.find((tab) => tab.key === "cfpOpen").current && (
-            <div className="bg-red-500 p-8 mt-12">
-              <h1>cfpOpen</h1>
-            </div>
-          )}
-
-        {tabs.find((tab) => tab.key === "inPerson") &&
-          tabs.find((tab) => tab.key === "inPerson").current && (
-            <div className="bg-red-500 p-8 mt-12">
-              <h1>inPerson</h1>
-            </div>
-          )}
-        {tabs.find((tab) => tab.key === "virtual") &&
-          tabs.find((tab) => tab.key === "virtual").current && (
-            <div className="bg-red-500 p-8 mt-12">
-              <h1>virtual</h1>
-            </div>
-          )} */}
       </Page>
     </>
-  );
-}
-
-export function UserTabs2({ tabs, setTabs, sortEvents, setSortEvents }) {
-  const classNames = (...classes) => classes.filter(Boolean).join(" ");
-  const changeTab = (e, value) => {
-    e.preventDefault();
-    console.log(e.target.value);
-    setTabs(
-      tabs.map((tab) =>
-        tab.title === e.target?.value || tab.title === value
-          ? { ...tab, current: true }
-          : { ...tab, current: false }
-      )
-    );
-  };
-
-  const getDataKeyAndSortKey = (tabName) => {
-    let dataKeyObj = {};
-    switch (tabName) {
-      case "cfpOpen":
-        dataKeyObj.dataKey = "cfpOpen";
-        dataKeyObj.sortKey = "date.start";
-        break;
-      case "inPerson":
-        dataKeyObj.dataKey = "inPerson";
-        dataKeyObj.sortKey = "date.start";
-        break;
-      case "virtual":
-        dataKeyObj.dataKey = "virtual";
-        dataKeyObj.sortKey = "date.start";
-        break;
-      default:
-        dataKeyObj.dataKey = "all";
-        dataKeyObj.sortKey = "date.start";
-    }
-    return dataKeyObj;
-  };
-
-  const sortUserTabItems = (tabName, order, events, setEvents) => {
-    const { dataKey, sortKey } = getDataKeyAndSortKey(tabName);
-    sortEvents[dataKey].sort(function (a, b) {
-      console.log(sortEvents[dataKey]);
-      const aVal = sortKey.includes(".")
-        ? getNested(a, sortKey.split("."))
-        : a[sortKey];
-      const bVal = sortKey.includes(".")
-        ? getNested(b, sortKey.split("."))
-        : b[sortKey];
-      if (tabName === "My Links") {
-        if (order === "ASC") {
-          return aVal.toLowerCase() > bVal.toLowerCase()
-            ? 1
-            : aVal.toLowerCase() < bVal.toLowerCase()
-            ? -1
-            : 0;
-        } else {
-          return aVal.toLowerCase() < bVal.toLowerCase()
-            ? 1
-            : aVal.toLowerCase() > bVal.toLowerCase()
-            ? -1
-            : 0;
-        }
-      } else {
-        if (order === "ASC") {
-          return new Date(aVal) > new Date(bVal)
-            ? 1
-            : new Date(aVal) < new Date(bVal)
-            ? -1
-            : 0;
-        } else {
-          return new Date(aVal) < new Date(bVal)
-            ? 1
-            : new Date(aVal) > new Date(bVal)
-            ? -1
-            : 0;
-        }
-      }
-    });
-    setSortEvents({ ...sortEvents });
-  };
-
-  const getNested = (obj, args) => {
-    return args.reduce((obj, level) => obj && obj[level], obj);
-  };
-
-  return (
-    <div>
-      <div className="sm:hidden">
-        <label htmlFor="tabs" className="sr-only">
-          Select a tab
-        </label>
-        <select
-          id="tabs"
-          name="tabs"
-          onChange={changeTab}
-          className="block w-full rounded-md border-gray-300 focus:border-indigo-500 focus:ring-indigo-500"
-          defaultValue={tabs.find((tab) => tab.current)?.title}
-        >
-          {tabs.map((tab) => (
-            <option key={tab.key}>{tab.title}</option>
-          ))}
-        </select>
-      </div>
-      <div className="hidden sm:block">
-        <div className="border-b border-gray-200">
-          <nav className="-mb-px flex" aria-label="Tabs">
-            {tabs.map((tab) => (
-              <a
-                key={tab.key}
-                href={tab.href}
-                onClick={(e) => changeTab(e, tab.title)}
-                className={classNames(
-                  tab.current
-                    ? "border-indigo-500 text-indigo-600"
-                    : "border-transparent text-gray-500 hover:text-gray-700 hover:border-gray-300",
-                  "w-1/4 py-4 px-1 text-center border-b-2 font-medium text-sm flex justify-center items-center gap-2"
-                )}
-                aria-current={tab.current ? "page" : undefined}
-              >
-                {tab.key === "all" && (
-                  <FaListUl title="Show All" className="text-base" />
-                )}
-                {tab.key === "virtual" && (
-                  <MdOutlineOnlinePrediction
-                    title="Virtual event"
-                    className="text-xl"
-                  />
-                )}
-                {tab.key === "inPerson" && (
-                  <MdOutlinePeople
-                    title="In person event"
-                    className="text-xl"
-                  />
-                )}
-                {tab.key === "cfpOpen" && (
-                  <FaMicrophoneAlt title="CFP is open" className="text-base" />
-                )}
-                <span>
-                  {tab.title} ({tab.total})
-                </span>
-                {tab.current && (
-                  <BiSortAlt2
-                    size="20"
-                    className="hover:text-gray-400"
-                    onClick={(e) => {
-                      e.preventDefault();
-                      e.stopPropagation();
-                      setTabs(
-                        tabs.map((tab) =>
-                          tab.current
-                            ? {
-                                ...tab,
-                                order: tab.order === "ASC" ? "DESC" : "ASC",
-                              }
-                            : { ...tab }
-                        )
-                      );
-                      sortUserTabItems(
-                        tab.title,
-                        tab.order === "ASC" ? "DESC" : "ASC"
-                      );
-                    }}
-                  />
-                )}
-              </a>
-            ))}
-          </nav>
-        </div>
-      </div>
-    </div>
   );
 }
